# api-docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "YukiBloom 网站的文章、图片、个人内容和站点页面 API。JSON 响应统一使用 code、message、data；成功 code 为 0，失败 code 与 HTTP 状态码一致。",
        "title": "YukiBloom 后端接口",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/admin/footprints": {
            "put": {
                "description": "需要登录和 CSRF 令牌；请求会替换全部地点、停留记录和路线。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理个人内容"
                ],
                "summary": "整体保存足迹",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录接口返回的 csrfToken",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "完整足迹数据",
                        "name": "footprints",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/personal.Footprints"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "保存后的足迹数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresponse.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/personal.Footprints"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "足迹数据无效",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "401": {
                        "description": "尚未登录",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "403": {
                        "description": "CSRF 校验失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "500": {
                        "description": "保存失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/timeline": {
            "put": {
                "description": "需要登录和 CSRF 令牌；请求会替换全部经历条目。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理个人内容"
                ],
                "summary": "整体保存实习经历",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录接口返回的 csrfToken",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "包含 items 实习经历数组",
                        "name": "timeline",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/personal.TimelineInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "保存后的时间线",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresponse.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "时间线数据无效",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "401": {
                        "description": "尚未登录",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "403": {
                        "description": "CSRF 校验失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "500": {
                        "description": "保存失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/footprints": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "个人内容"
                ],
                "summary": "查询公开足迹",
                "responses": {
                    "200": {
                        "description": "地点、停留记录和路线",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresponse.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/personal.Footprints"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            }
        },
...
        "/api/v1/timeline": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "个人内容"
                ],
                "summary": "查询公开实习经历",
                "responses": {
                    "200": {
                        "description": "包含 items 数组的时间线",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresponse.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apiresponse.Envelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "personal.Footprints": {
            "type": "object",
            "properties": {
                "locations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/personal.Location"
                    }
                },
                "routes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/personal.Route"
                    }
                },
                "stays": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/personal.Stay"
                    }
                }
            }
        },
        "personal.Internship": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "iconColor": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isPresent": {
                    "type": "boolean"
                },
                "position": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "integer"
                },
                "startDate": {
                    "type": "string"
                }
            }
        },
        "personal.Location": {
            "type": "object",
            "properties": {
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "personal.Route": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "label": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "integer"
                },
                "to": {
                    "type": "string"
                },
                "transport": {
                    "type": "string"
                }
            }
        },
        "personal.Stay": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isPresent": {
                    "type": "boolean"
                },
                "locationId": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "integer"
                },
                "startDate": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "personal.TimelineInput": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/personal.Internship"
                    }
                }
            }
        },
...
    }
}